WIP: 2c-d cleanup — 428 errors remaining (ServerOutputId unified to events; decode_agent_response_event returns Result; total_workspace_bonus_credits_remaining returns i32; last_modified_at returns DateTime<Local>; get_preceding_user_query callers use as_deref/unwrap_or)

+++ app/src/settings_view/billing_and_usage_page.rs
@@ -45,8 +45,8 @@ impl AIRequestUsageModel {
     pub fn next_refresh_time(&self) -> Option<chrono::DateTime<chrono::Utc>> { None }
     pub fn request_limit(&self) -> usize { 0 }
     pub fn requests_used(&self) -> usize { 0 }
-    pub fn total_workspace_bonus_credits_remaining<W>(&self, _: W) -> i64 { 0 }
-    pub fn total_current_workspace_bonus_credits_remaining<W>(&self, _: W) -> i64 { 0 }
+    pub fn total_workspace_bonus_credits_remaining<W>(&self, _: W) -> i32 { 0 }
+    pub fn total_current_workspace_bonus_credits_remaining<W>(&self, _: W) -> i32 { 0 }
     pub fn compute_buy_addon_credits_banner_display_state<W>(&self, _: W) -> crate::terminal::buy_credits_banner::BuyCreditsBannerDisplayState { crate::terminal::buy_credits_banner::BuyCreditsBannerDisplayState::Hidden }
 }
 use crate::{
